:sparkles: search emojis, support utf8 symbol, global .czrc config from  (#14), close issues
* 🐛 Fix exception if no package.json is present

If no package.json could be found cz-emoji would throw an exception when
trying to access res.pkg.config.

* ✨ Use autocomplete with fuzzy search

Use autocomplete with fuzzy search to select the emoji rather than
scrolling forever.

* ✨ Add option 'symbol' to insert the unicode symbols directly

Setting symbol to true inserts the unicode symbols directly, instead of adding
a text representation.

* ✨ Search the .czrc for cz-emoji options

* ✨ Add closed issues to the commit message

This part seems to have been forgotten. The prompt asks for the information but it was never used.