📝 (Readme/Customization) Adding Symbol Config
Adding documentation for a configuration feature that was previously not shared. In the config there
is a setting that allows for the unicode value (`types.emoji`) to be used rather than the gitmoji
markup (`types.code`). In some cases it is preferred to use the unicode character rather than the
gitmoji when it comes to performing a copy and paste of commit messages in some environments.