:bug: don't duplicate subject prefix in the final output
Counting the length of the whole `type`, `scope` and `subject` combo using `maxlength-input` has a side-effect.
It prepends `subject` by the formatted `type` and `scope` before we reach the `format` stage.
As a consequence in `format` we were formatting it again and ended up having a duplicated prefix (cf. previous commit messages).

This commit assumes that `subject` is already formatted with `type` and `scope` when reaching the `format` stage.
It makes the code a little less clearer as now it's not very obvious that the formatting is happening during the prompt phase.
However I think that's fine as it offers a nice feature with minimal changes to the code ¯\_(ツ)_/¯